requirements.txt


Flask
gunicorn
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
